add ShopSection component and integrate into Menswear, Womenswear, and Children pages
