+++ client/src/pages/Womenswear/Womenswear.js
@@ -2,4 +2,6 @@ import React from 'react';
 import { ShopSection } from '../../components/organisms/ShopSection/ShopSection';
 import { sectionConfigs } from '../../utils/constants';
 
-export const Children = () => <ShopSection config={sectionConfigs.children} />;
+export const Womenswear = () => (
+  <ShopSection config={sectionConfigs.womenswear} />
+);
