+++ client/src/pages/Womenswear/index.js
@@ -1,0 +1,1 @@
+export { Womenswear } from './Womenswear';
